requirements.txt
django
djangorestframework
django-cors-headers
channels
daphne
graphene-django
django-graphql-jwt
psycopg2-binary